rename variationals to models
