+++ edward/__init__.py
@@ -1,14 +1,12 @@
 from __future__ import absolute_import
+from . import models
 from . import stats
 from . import data
 from . import inferences
-from . import models
 from . import util
-from . import variationals
 
 # Direct imports for convenience
+from .models import *
 from .data import *
 from .inferences import *
-from .variationals import *
-from .models import PythonModel, StanModel, PyMC3Model
 from .util import set_seed
