+++ edward/models/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import
+from .models import *
+from .variationals import *
