+++ edward/models/variationals.py
