+++ examples/bernoulli.py
@@ -8,8 +8,8 @@ Variational model
 import edward as ed
 import tensorflow as tf
 
+from edward.models import Variational, Bernoulli
 from edward.stats import bernoulli
-from edward.variationals import Variational, Bernoulli
 
 class BernoulliPosterior:
     """
