+++ examples/beta_bernoulli_stan.py
@@ -9,7 +9,7 @@ Variational model
     Likelihood: Mean-field Beta
 """
 import edward as ed
-from edward.variationals import Variational, Beta
+from edward.models import Variational, Beta
 
 model_code = """
     data {
