+++ examples/hierarchical_logistic_regression.py
@@ -14,8 +14,8 @@ import tensorflow as tf
 import matplotlib.pyplot as plt
 import numpy as np
 
+from edward.models import Variational, Normal
 from edward.stats import bernoulli, norm
-from edward.variationals import Variational, Normal
 
 class HierarchicalLogistic:
     """
